pyproject.toml
[project]
name = "ocr-poc"
version = "0.1.0"
description = "Proof of Concept for comparing different OCR technologies"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "OCR POC Team"}
]

dependencies = [
    # Core dependencies
    "numpy>=1.25.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    
    # Utility libraries
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "scikit-learn>=1.3.0",
    
    # Base OCR libraries (will be expanded after research)
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_poc"]

[tool.hatch.build.targets.sdist]
include = [
    "src/ocr_poc",
    "tests",
    "README.md",
    "LICENSE",
]
